% Test test

\name{compute}
\title{Compute a number}
\usage{
compute(x)
}
\arguments{
  \item{x}{Any number}
}
\description{
  \code{compute} is a function that always returns 22
}
\details{
  
}
\value{
  22
}
\references{
  Chambers, J. M. and Hastie, T. J. (1992)
  \emph{Statistical Models in S}.
  Wadsworth & Brooks/Cole.
}
\examples{
stopifnot(compute(1) == 22)
}
\keyword{regression}
\keyword{models}
